+++ drift_dev/lib/src/cli/commands/schema/generate_utils.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 import 'package:args/command_runner.dart';
 import 'package:dart_style/dart_style.dart';
 import 'package:path/path.dart' as p;
+import 'package:collection/collection.dart';
 
 import '../../../analysis/results/file_results.dart';
 import '../../../analysis/results/results.dart';
@@ -66,23 +67,34 @@ class GenerateUtilsCommand extends Command {
       final version = versionAndEntities.key;
       final entities = versionAndEntities.value;
 
-      await _writeSchemaFile(
-        outputDir,
-        version,
-        entities,
-        argResults?['data-classes'] as bool,
-        argResults?['companions'] as bool,
-      );
+      final file = File(
+          p.join(outputDir.path, GenerateUtils._filenameForVersion(version)));
+      await file.writeAsString(GenerateUtils.generateSchemaCode(
+          cli,
+          version,
+          entities,
+          argResults?['data-classes'] as bool,
+          argResults?['companions'] as bool));
     }
 
     final versions = schema.keys.toList()..sort();
-    await _writeLibraryFile(outputDir, versions);
+    final libraryFile = File(p.join(outputDir.path, 'schema.dart'));
+    await libraryFile
+        .writeAsString(GenerateUtils.generateLibraryCode(versions));
     print(
         'Wrote ${schema.length + 1} files into ${p.relative(outputDir.path)}');
   }
+}
 
-  Future<void> _writeSchemaFile(
-    Directory output,
+class GenerateUtils {
+  static String _filenameForVersion(int version) => 'schema_v$version.dart';
+  static const _prefix = '// GENERATED CODE, DO NOT EDIT BY HAND.\n'
+      '// ignore_for_file: type=lint';
+  static final _dartfmt = DartFormatter();
+
+  /// Generates Dart code for a specific schema version.
+  static String generateSchemaCode(
+    DriftDevCli cli,
     int version,
     ExportedSchema schema,
     bool dataClasses,
@@ -105,7 +117,6 @@ class GenerateUtilsCommand extends Command {
         imports: NullImportManager(),
       ),
     );
-    final file = File(p.join(output.path, _filenameForVersion(version)));
 
     writer.leaf()
       ..writeln(_prefix)
@@ -126,10 +137,12 @@ class GenerateUtilsCommand extends Command {
 
     DatabaseWriter(input, writer.child()).write();
 
-    return file.writeAsString(_dartfmt.format(writer.writeGenerated()));
+    return _dartfmt.format(writer.writeGenerated());
   }
 
-  Future<void> _writeLibraryFile(Directory output, Iterable<int> versions) {
+  /// Generates the Dart code for a library file that instantiates the schema
+  /// for each version.
+  static String generateLibraryCode(Iterable<int> versions) {
     final buffer = StringBuffer()
       ..writeln(_prefix)
       ..writeln('//@dart=2.12')
@@ -153,19 +166,17 @@ class GenerateUtilsCommand extends Command {
         ..writeln('return v$version.DatabaseAtV$version(db);');
     }
 
-    final missingAsSet = '{${versions.join(', ')}}';
+    final versionsSet =
+        '[${versions.sorted((a, b) => a.compareTo(b)).join(', ')}]';
     buffer
       ..writeln('default:')
-      ..writeln('throw MissingSchemaException(version, const $missingAsSet);')
-      ..writeln('}}}');
+      ..writeln('throw MissingSchemaException(version, versions);')
+      ..writeln('}}');
 
-    final file = File(p.join(output.path, 'schema.dart'));
-    return file.writeAsString(_dartfmt.format(buffer.toString()));
+    buffer
+      ..writeln('static const versions = const $versionsSet;')
+      ..writeln('}');
+
+    return _dartfmt.format(buffer.toString());
   }
-
-  String _filenameForVersion(int version) => 'schema_v$version.dart';
-
-  static final _dartfmt = DartFormatter();
-  static const _prefix = '// GENERATED CODE, DO NOT EDIT BY HAND.\n'
-      '// ignore_for_file: type=lint';
 }